Add compatibility with migrations framework
There is no longer a `sync_db` signal since the switch to the
migrations framework. Instead, we need to use the `post_migrate`
signal.

Note that this change has also been applied upstream (in v1.3.2)
but that build also introduces at least one bug which causes
django-migrations to break for general use. v1.3.1 does not
contain this change, so we need to apply it here, and cannot
currently upgrade to 1.3.2